# domain-product-pricing-core/src/main/java/com/firefly/domain/product/pricing/core/fees/commands/RegisterFeeApplicationRuleCommand.java
/*
 * Copyright 2025 Firefly Software Solutions Inc
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.firefly.domain.product.pricing.core.fees.commands;

import com.firefly.common.cqrs.command.Command;
import com.firefly.common.product.sdk.model.FeeApplicationRuleDTO;
import lombok.Data;
import lombok.EqualsAndHashCode;

import java.util.UUID;

@EqualsAndHashCode(callSuper = true)
@Data
public class RegisterFeeApplicationRuleCommand extends FeeApplicationRuleDTO implements Command<UUID> {
    private UUID feeComponentId;

    public RegisterFeeApplicationRuleCommand withFeeComponentId(UUID feeComponentId) {
        this.setFeeComponentId(feeComponentId);
        return this;
    }
}